feat: added Bludge Town and adjusted layout
